requirements.txt
keras==2.15.0
numpy==1.24.3
Pillow==10.1.0
scikit_learn==1.3.0
streamlit==1.29.0
tensorflow==2.15.0
